Fix bug, which lets images snap back to their original location and camera angle after modifying them
The bug was introduced with the recent change of the image paint style in 34d83387d217c396c76c3d5eedaadc5afa88edb6. From there on MapillaryAbstractImage#getLatLon() was used instead of MapillaryAbstractImage#getMovingLatLon() and the same with the camera angle.
This is now fixed by simply using the getMovingLatLon() and getMovingCa() methods instead.