Fix NullPointerException in SelectMode.mouseMoved()
Previously in the mouseMoved() method of SelectMode, it was checked, if the data of the current edit layer was null. But without checking, if the edit layer itself was null.
Now, there is such a check in place, that skips this part, when there is no edit layer at the moment.
This issue has been reported as https://josm.openstreetmap.de/ticket/13624