Add new source set, which is compiled against the minimum required JOSM version
This should alert us, if we introduce changes that require to bump the minimum required version. This could be necessary, if a part of the JOSM API is used, that was introduced in a later JOSM version.

The task 'minJosmVersionCompile' is executed in the Travis build and fails, when compilation against the minimum required version fails.