Make license choice clearer and distribute markdown texts as HTML
The GPLv2 license is now only named `LICENSE` to make it clear that this is our main license.
The file `LICENSE.md` contains the more detailed explanation icluding the exceptions and now the correct copyright notice.
Also the files `README.md` and `LICENSE.md` will now be distributed as HTML files instead of plain markdown for better readability.