Require the oraclejdk9 build on Travis to pass
This build was allowed to fail, because JaCoCo 0.7.7 caused it to fail.
The issue is fixed in the current snapshot of JaCoCo 0.7.8, so we'll use this until 0.7.8 is released.
That means that Travis CI can now require again, that the Java 9 build passes.

Another small change is, that `targetJdk` is now omitted for the PMD configuration. gradle ignores this property for PMD versions >= 5.0, so it is not needed.