Add link icon and icons for info panel
The icons will soon be used for hyperlinking buttons and for a new panel which displays meta information for the displayed mapillary image.
Also updated license choice to Apache&SIL instead of CC-BY&SIL, that way the license text(s) is/are much shorter and not even the icons had to be switched.