Add new panel for displaying metadata for the selected Mapillary image
This also incorporates the actions that were previously hidden in the HypertextLabel ("open image in browser", "copy image key" and also "copy mapillary tag", which is now "tag selected object with mapillary tag").
Currently it basically just shows the image key and sequence key, but could be extended to show more metadata in the future.