Add a popup for notifying users of the new ImageInfoPanel
The popup is displayed, whenever a user hovers the first Mapillary image in a JOSM session.
It hides for the rest of the session, when the user clicks anywhere outside the popup.
It can also be hidden indefinitely by clicking "I got it, close this.", or by opening the ImageInfoPanel once.
Otherwise the popup will at maximum appear four times for one given installation, before it disappears forever.