Separate the URLs for APIv3 and the ones for the main website from the ones for APIv2
The URL-generating methods for APIv3 and the main website are now inside static inner classes called `APIv3` and `MainWebsite`. This separates them from each other and allows the developers to instantly see, which API is used for a call (`MapillaryURL.APIv3.searchSequences()` or `APIv3.searchSequences()` vs. `MapillaryURL.searchSequences()`)
A new method (APIv3.parseNextFromLinkHeader()) is introduced. It allows for extraction of the URL for the next page from the Link header, as explained in https://www.mapillary.com/developer/api-documentation/#pagination .
Also the method searchSequences(Bounds) is now also implemented for APIv3.
The changes are accompanied by tests for the newly created methods.