Improve icon download mechanism for map objects
The getIcon() method now has two operation modes (toggled by a boolean parameter).
In one of them, only the MapObject instance and the cache are searched for the icon. The other one will open a connection to a URL and download the icon from there.
The former one is for cases when the method needs to return fast (e.g. when painting a Graphics instance), the latter one can be called when it's ok that the method blocks while downloading the icon.