Fix repainting issue by invalidating the MapillaryLayer correctly
The issue was, that the MapillaryLayer was only repainted when the map was panned or zoomed. It occured with any version of the plugin and only with JOSM versions 12170 or higher.

It was caused by the change in https://josm.openstreetmap.de/changeset/12170/josm , which caused that from then on only these layers are repainted, which have been invalidated.
The reasoning behind that change in JOSM is explained in https://josm.openstreetmap.de/ticket/13175 .

Previously any change that required a repaint fired a complete repaint() of the MapView, because the MapillaryLayer did not implement the method isChanged(), so MapillaryLayer.isChanged() always returned false and the layer was repainted on every repaint() of the MapView.
Because the method isChanged() now returns false by default and it is no longer considered when deciding which layers should be repainted, I replaced all uses of MapillaryData.dataUpdated() (which internally called MapView.repaint()) by calls to the method MapillaryLayer.invalidate().

Fixes #59.