Add test case for decoding a user profile that is JSON encoded
This test case covers what happens when an avatar points to a URL that is not an image, or at least can't be read by ImageIO.read().
Also, the last missing bits of the package `org.openstreetmap.josm.plugins.mapillary.utils.api` have received test coverage, we are now at 100% for this package.