Properly reset the layers when testing
When resetting the state of the LayerManager, make sure to also remove the MapillaryLayer as an ActiveLayerChangeListener (e.g. by destroy()ing it). Otherwise you might run into problems when you later want to destroy() the MapillaryLayer and the LayerManager complains that the MapillaryLayer can't be removed as an ActiveLayerChangeListener, because it has already been removed before by the resetState() method.