Fix build by updating JOSM version to compile against
The build was broken (again) by an update to the artifact https://josm.openstreetmap.de/download/josm-unittest.jar , which now relies on classes, that are only available in JOSM versions 12866 or later.
As part of the new JOSM version, the TestUtil now needs to initialize the class `Config`, by setting the preferences instance and the base directory provider.