Catch IllegalArgumentException "Layer MapillaryLayer is not managed by us."
See https://josm.openstreetmap.de/ticket/15257 . Probably this had to do with concurrent modification of the layers (e.g. MapillaryLayer was removed while trying to activating it).
Also replace the `DelayedDownload` class with a lambda that downloads immediately. Why delay, when it's submitted to a worker anyway? Also 1.5 seconds seemed somewhat arbitrary.