Don't throw IllegalStateException when downloading in "wrong" mode
Calling the methods MapillaryDownloader.downloadOSMArea() or MapillaryDownloader.downloadVisibleArea() in the "wrong" mode doesn't hurt, so no need to throw an exception.
This commit also changes, how `MapillaryDownloadViewAction` determines when it's active. It now listens for changes in the current download mode and updates its state when needed.

Fixes https://josm.openstreetmap.de/ticket/14981 .