Make the "download area too big" messages less obtrusive, show them only once and as Notification
The idea to make this a Notification came from https://josm.openstreetmap.de/ticket/15220 . In https://josm.openstreetmap.de/ticket/13389 and https://josm.openstreetmap.de/ticket/15024 it was also noted that the previous message was showing too often and it wasn't really clear what it referred to. It was e.g. not obvious that the Mapillary plugin was the source of that message.
This also gets rid of the "tempSemiautomatic" download mode and improves interrupting behaviour of the download tasks.