Fix unit tests that were timing out
The two fixed unit tests that timed out were related to URLs that were supposed to be invalid/non-existent. For that purpose I used domains like https://invalidURL, but that seems to lead to timeouts. I now changed that to use https://example.org, which should work equally well and not time out.