Open Mapillary layer when importing images from geotagged images layer
That was an oversight in c8f1ec0c6d64ae44c81cbe1836518920f3cea7e6. The images were added to the Mapillary layer but the layer was not displayed. You had to open the layer manually, starting with this commit it opens automatically again.

Also, the "Import" button when selecting the "geotagged images" layer is now disabled as long as there are no layers selected.