Better parallelism for getting additional image information
This converts from Threads to ForkJoinPool for downloads. This doesn't make a
major difference (yet), since it currently takes ~3.5 seconds to get the
headers for the request, and roughly the same amount of time to get the
full response.

That being said, ~200 seconds was shaved off of the time in the small
sample area.

Signed-off-by: Taylor Smock <taylor.smock@kaart.com>