Only get traffic sign detections for all images and get image detections on demand
This means we are only downloading full detections on object selection,
so this reduces load on Mapillary's servers and reduces memory usage in
JOSM (not as many detections are kept in memory).

The major downside is that it takes a few seconds after the download
starts to get all the relevant detections.

Signed-off-by: Taylor Smock <taylor.smock@kaart.com>