FIXUP: Error when trying to upload images
This fixes #102.

The root issue is that the ArrayBlockingQueue was initialized with a
size of 15, and the network was not fast enough to upload images prior
to filling up the queue and then erroring out.

The fix is using the size of the images set to set the maximum size of
the queue.

Signed-off-by: Taylor Smock <tsmock@fb.com>