FIXUP: Fix JOSM-20211
The crash occurs when Mapillary is set to cut off sequences at download
bounds. What happened is that the new bounds were used to determine what
was in the sequences downloaded in the new bounds, so if a sequence was
already downloaded in a different bounding box, this error occurred.

Signed-off-by: Taylor Smock <tsmock@fb.com>