Fix a download issue
The download issue caused sequences to not show up, since it was usually
the first page (with most of the sequence information) that failed to
download. We now try one more time before giving up. Usually, this
works. This also applies to other code that reuses the same
BoundsDownloadRunnable class.

The triggers for the retry are:
* Server response code of 504
* We have not yet hit the maximum number of retries.

Also fix an issue where images could not be set to visible after
becoming invisible.