Fix JOSM #21070 and #21072 (hopefully)
The issue occurred when there were two objects in the stream with the
same key. I don't know how it happened, and was unable to reproduce even
when adding a second key into the json.

This patch hopefully fixes the issue, but due to not being able to
reproduce, it is unknown if this fixes the issue.

Signed-off-by: Taylor Smock <tsmock@fb.com>