Import: Drop image imports
Since Mapillary does not have a public upload API, and we now use JOSM's
image viewer, it is pointless for users to "import" images into the
Mapillary layer. JOSM also supports opening directories of images now.

In the event upload is re-added, it will most likely be on a GeoImage by
GeoImage layer basis, and will not import the images into the Mapillary
layer.

Signed-off-by: Taylor Smock <tsmock@fb.com>