MapillaryLayer: Improve rendering performance
Fix an expensive recurrent call by storing initial results and passing
to required locations.

getSelectedNodes took up 60%+ of time spent in drawImageMarker, which
took up ~75% of the time in paintWithLock. This means getSelectedNodes
took up ~46% of the processing time in the paintWithLock method.

getSelectedNodes now takes up 0.15% of the time in the paintWithLock
method.

Signed-off-by: Taylor Smock <tsmock@fb.com>