Fix #21121: Mapillary unit tests are failing in JOSM Jenkins
This appears to be caused due to the test resources directory being set
in the JOSM `/plugins` directory instead of the Mapillary plugin
directory (`/plugins/Mapillary`).

This works around the problem by checking to see if the `test/resources`
directory exists, then if it does not, it falls back to `Mapillary-git`
or `Mapillary` (in that order) with the `test/resources` appended.

Signed-off-by: Taylor Smock <tsmock@fb.com>